+++ js/temp-data.js
@@ -1,0 +1,59 @@
+const MIN_NUMBER_AVATAR_IMG = 1;
+const MAX_NUMBER_AVATAR_IMG = 8;
+
+const X_MIN = 35.65000;
+const X_MAX = 35.70000;
+const Y_MIN = 139.70000;
+const Y_MAX = 139.80000;
+const MIN_PRICE = 0;
+const MAX_PRICE = 1000000;
+const MIN_ROOMS = 1;
+const MAX_ROOMS = 100;
+const MIN_GUESTS = 1;
+
+const TITLES = [
+  'Солнечное место',
+  'В самом сердце города',
+  'Уютное местечко',
+  'У театра',
+  'Комфорт класс',
+];
+const TYPES = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+];
+const CHECKINS = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+const CHECKOUTS = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+const DESCRIPTIONS = [
+  'Незабываемый отдых',
+  'Свежий ремонт, без окон, в центре города',
+  'С отличным ремонтом',
+  'Рядом расположены большой торговый центр, кинотеатры, аквапарк',
+  'Мы сделали все, чтобы наши гости чувствовали уют и комфорт ',
+];
+const PHOTOS = [
+  'http://o0.github.io/assets/images/tokyo/hotel1.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel2.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel3.jpg',
+];
+
+export  {MIN_NUMBER_AVATAR_IMG, MAX_NUMBER_AVATAR_IMG, X_MIN, X_MAX, Y_MIN, Y_MAX, MIN_PRICE, MAX_PRICE, MIN_ROOMS,
+  MAX_ROOMS, MIN_GUESTS, TITLES, TYPES, CHECKINS, CHECKOUTS, FEATURES, DESCRIPTIONS, PHOTOS};
